+++ libs/nest/prisma/migrations/20241120194124_add_weight_and_emission_factors_to_product_materials_and_materials/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "materials" ADD COLUMN     "emissions_factor" DOUBLE PRECISION;
+
+-- AlterTable
+ALTER TABLE "product_materials" ADD COLUMN     "weight" DOUBLE PRECISION;