+++ libs/react/src/interfaces/products.ts
@@ -17,11 +17,13 @@ export interface ProductsQuery {
     id: string
     yield: number | null;
     unitOfMeasure: string | null;
+    weight: number | null;
     material: {
       id: string
       name: string
       materialCategory: string | null;
       materialSubcategory: string | null;
+      emissionsFactor: number | null;
       attributeAssurances: EntityLevelAttributeAssuranceGraphQL[];
       materialSuppliers: {
         id: string;